Fixed build warnings on MSVC14 x64 in the XGL loader.

+++ code/XGLLoader.cpp
@@ -711,7 +711,7 @@ unsigned int XGLImporter::ResolveMaterialRef(TempScope& scope)
     const std::string& s = GetElementName();
     if (s == "mat") {
         ReadMaterial(scope);
-        return scope.materials_linear.size()-1;
+        return static_cast<unsigned int>(scope.materials_linear.size()-1);
     }
 
     const int id = ReadIndexFromText();
